added double data types

+++ thesis/Core/ViewModel/TotalWeightViewModel.cs
@@ -5,18 +5,18 @@ namespace thesis.Core.ViewModel
     public class TotalWeightViewModel
     {
         //Dashboard
-        public int TotalWeight { get; set; }
-        public int DailyWeight { get; set; }
-        public int WeeklyWeight { get; set; }
-        public int MonthlyWeight { get; set; }
-        public int YearlyWeight { get; set; }
+        public double TotalWeight { get; set; }
+        public double DailyWeight { get; set; }
+        public double WeeklyWeight { get; set; }
+        public double MonthlyWeight { get; set; }
+        public double YearlyWeight { get; set; }
         // for bar data
-        public List<int> Suspect { get; set; }
-        public List<int> Condemned { get; set; }
-        public List<int> Pass { get; set; }
+        public List<double> Suspect { get; set; }
+        public List<double> Condemned { get; set; }
+        public List<double> Pass { get; set; }
         // for area chart
-        public List<int> monthlyRangeApproved { get; set; }
-		public List<int> monthlyRangeCondemned { get; set; }
+        public List<double> monthlyRangeApproved { get; set; }
+		public List<double> monthlyRangeCondemned { get; set; }
 		public string[] monthAbbreviationsArray { get; set; }
 	}
 }
